requirements.txt
Jinja2>=3.1.2
kubernetes>=24.2.0,<=27.2.0
PyYAML>=6.0.0,<=6.0.1
requests==2.31.0
tabulate>=0.8.9,<=0.9.0
termcolor<3.0
urllib3>=1.26.8,<=2.0.4
